core_cli/requirements.txt
# ============================================================================
# Flynt Studio - Backend Requirements
# ============================================================================

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Utilities
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Web Server & Async
starlette>=0.27.0
httpx>=0.25.0
aiofiles>=23.2.0

# LLM Providers (install as needed)
requests>=2.31.0  # For OpenRouter API calls
ollama>=0.1.0  # For Ollama local LLM

# Database & Data
sqlalchemy>=2.0.0
sqlite3-python>=1.0.0  # SQLite support
alembic>=1.13.0  # Database migrations

# API & HTTP
requests>=2.31.0
urllib3>=2.0.0

# ML & Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Data Validation & Serialization
marshmallow>=3.20.0
jsonschema>=4.19.0

# Logging & Monitoring
python-json-logger>=2.0.7
structlog>=23.2.0

# Async Support
asyncio-contextmanager>=1.0.0

# Error Handling & Resilience
tenacity>=8.2.0  # Sophisticated retry logic

# Monitoring & Observability
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
responses>=0.23.0  # Mock HTTP responses
unittest-mock>=1.5.0

# Code Quality (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation (optional)
pdocs>=1.0.0

pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing async endpoints

# Code Quality (optional, for development)
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0

# Optional: Advanced Agent Orchestration (for future expansion)
# crewai>=0.1.0
# langchain>=0.1.0
# anthropic>=0.7.0

# Optional: Additional LLM Providers
# openai>=1.3.0  # For GPT models
# cohere>=4.37.0  # For Cohere models
# replicate>=0.15.0  # For Replicate models

# Optional: Vector Search & RAG
# chromadb>=0.4.0  # Vector database
# pinecone-client>=2.2.0  # Pinecone vector DB
# weaviate-client>=4.0.0  # Weaviate vector DB
# pgvector>=0.2.0  # PostgreSQL vector extension

# Optional: Document Processing
 pypdf>=3.17.0
 python-pptx>=0.6.21
 python-docx>=0.8.11
 pillow>=10.0.0